Add files via upload
Write a function that takes 1 arg, a string containing the 3 character abbreviation for a month name. Return an integer that represents that month. The integer should be the same month value that's used by the Date constructor to create a Date object. For example, the integer that represents "Jun" is 5.